Add Model choice UI (#726)
* added ui to select both models and backbones

* reorganized retinanet backbones imports

* added missing import

* make widgets visible to other methods

* cleaned up model lists

* move resnet to be shared by all torchvision models

* import from common resnet_fpn module

* import backbones from faster_rcnn folder

* import backbones from keypoint_rcnn folder

* import backbones from mask_rcnn folder

* fixed backbones reference

* added quickstart_object_detection notebook

* pip install icevision from master

* specific to retinanet model

* backbones specific to non-retinanet models

* applied black

* move param_groups to resnet_fpn_utils

* added mask_rcnn to mmdet model list

* re-organize mmdet mask_rcnn like retinanet
