+++ icevision/backbones/__init__.py
@@ -1,6 +1,7 @@
 from icevision.backbones.vgg import *
 from icevision.backbones.resnet import *
 from icevision.backbones.resnet_fpn import *
+from icevision.backbones.resnet_fpn_utils import *
 from icevision.backbones.resnest_fpn import *
 from icevision.backbones.mobilenet import *
