+++ icevision/models/mmdet/models/mask_rcnn/__init__.py
@@ -1,6 +1,6 @@
+from icevision.models.mmdet.models.mask_rcnn import backbones
 from icevision.models.mmdet.common.mask.two_stage import *
-from icevision.models.mmdet.models.mask_rcnn.backbones import *
-from icevision.models.interpretation import Interpretation
+from icevision.models.interpretation import Interpretation, _move_to_device
 from icevision.models.mmdet.common.interpretation_utils import (
     sum_losses_mmdet,
     loop_mmdet,
