+++ icevision/models/mmdet/models/mask_rcnn/backbones/__init__.py
@@ -1,0 +1,1 @@
+from icevision.models.mmdet.models.mask_rcnn.backbones.resnet_fpn import *
