+++ icevision/models/mmdet/models/retinanet/backbones/__init__.py
@@ -1,0 +1,1 @@
+from icevision.models.mmdet.models.retinanet.backbones.resnet_fpn import *
