+++ icevision/models/torchvision/keypoint_rcnn/__init__.py
@@ -1,6 +1,6 @@
-from icevision.models.torchvision import backbones
 from icevision.models.torchvision.loss_fn import *
 
+from icevision.models.torchvision.keypoint_rcnn import backbones
 from icevision.models.torchvision.keypoint_rcnn.dataloaders import *
 from icevision.models.torchvision.keypoint_rcnn.model import *
