+++ icevision/models/torchvision/keypoint_rcnn/backbones/__init__.py
@@ -1,0 +1,1 @@
+from icevision.backbones.resnet_fpn import *
