+++ icevision/models/torchvision/mask_rcnn/__init__.py
@@ -1,7 +1,6 @@
-# import icevision.models.torchvision.backbones as backbones
-from icevision.models.torchvision import backbones
 from icevision.models.torchvision.loss_fn import *
 
+from icevision.models.torchvision.mask_rcnn import backbones
 from icevision.models.torchvision.mask_rcnn.dataloaders import *
 from icevision.models.torchvision.mask_rcnn.model import *
 from icevision.models.torchvision.mask_rcnn.prediction import *
