+++ tests/models/torchvision_models/retinanet/conftest.py
@@ -4,5 +4,5 @@ from icevision.all import *
 
 @pytest.fixture
 def fridge_retinanet_model() -> nn.Module:
-    backbone = retinanet.backbones.resnet_fpn.resnet18_fpn(pretrained=False)
+    backbone = models.torchvision.retinanet.backbones.resnet18_fpn(pretrained=False)
     return retinanet.model(num_classes=5, backbone=backbone)
